refactor navbar, styling
