+++ paws-app/components/Layout/index.js
@@ -1,10 +1,14 @@
 import Navbar from "./components/Navbar";
-
+import { Container } from "@mantine/core";
 export default function Layout ({children}) {
-    return (
-        <>
-        <Navbar/>
-        <main>{children}</main>
-        </>
-    )
+  return (
+    <>
+      <Navbar />
+      <main>
+        <Container my={20}>
+          {children}
+        </Container>
+      </main>
+    </>
+  )
 }
