+++ paws-app/pages/mobile/index.js
@@ -1,46 +1,42 @@
 import Image from 'next/image';
 import Link from 'next/link';
-
 import { Center, Button, Text, Title } from '@mantine/core';
 import styles from '../../styles/Mobile.module.css';
-// import { useDispatch } from 'react-redux';
-// import { setNavbarTitle } from '../../components/shared/reducers/navbar';
 
 export default function MobileLandingPage() {
-    // const dispatch = useDispatch();
-    return (
+  return (
     <>
-        <div className={`${styles.topSpacer}`}>
-            <Center>
-                <div className={`${styles.logoContainer}`}>
-                    <Image src={'/paws_logo.png'} width={280} height={280} />
-                </div>
-            </Center>
-            <Center>
-                <Title order={3}><Text transform='uppercase' weight={700} align='center'>Welcome to PAWS</Text></Title>
-            </Center>
-            <Center>
-                <div className={styles.textSubheader}>
-                    <Text weight={500} align='center'>Learn more about our shelter animals with a quick qr code scan.</Text>
-                </div>
-            </Center>
-        </div>
+      <div className={`${styles.topSpacer}`}>
         <Center>
-            <div className={`${styles.buttonSpacer}`}>
-                <Link href='/mobile/scan'>
-                    <Button 
-                        radius='xl' 
-                        size='lg' 
-                        fullWidth 
-                        uppercase 
-                        style={{ backgroundColor: '#458642'}}
-                        onClick={() =>{
-                            // dispatch(setNavbarTitle('QR Scan'))
-                        }}
-                    >Scan Now</Button>
-                </Link>
-            </div>
+          <div className={`${styles.logoContainer}`}>
+            <Image src={'/paws_logo.png'} width={280} height={280} />
+          </div>
         </Center>
+        <Center>
+          <Title order={3}><Text transform='uppercase' weight={700} align='center'>Welcome to PAWS</Text></Title>
+        </Center>
+        <Center>
+          <div className={styles.textSubheader}>
+            <Text weight={500} align='center'>Learn more about our shelter animals with a quick qr code scan.</Text>
+          </div>
+        </Center>
+      </div>
+      <Center>
+        <div className={`${styles.buttonSpacer}`}>
+          <Link href='/mobile/scan'>
+            <Button 
+              radius='xl' 
+              size='lg' 
+              fullWidth 
+              uppercase 
+              style={{ backgroundColor: '#458642'}}
+              onClick={() =>{
+                  // dispatch(setNavbarTitle('QR Scan'))
+              }}
+            >Scan Now</Button>
+          </Link>
+        </div>
+      </Center>
     </>
-    );
+  );
 }
